+++ src/command/args.rs
@@ -23,8 +23,7 @@ pub struct EncodeToOutput {
     pub output: Option<PathBuf>,
 
     /// Set the output ffmpeg audio codec.
-    /// By default when the input & output file extension match 'copy' is used,
-    /// otherwise 'libopus'.
+    /// By default 'copy' is used. Otherwise, if re-encoding is necessary, 'libopus' is default.
     ///
     /// See https://ffmpeg.org/ffmpeg.html#Audio-Options.
     #[arg(long = "acodec")]
@@ -36,6 +35,12 @@ pub struct EncodeToOutput {
     /// No effect if the input audio has 3 or fewer channels.
     #[arg(long)]
     pub downmix_to_stereo: bool,
+
+    /// Only process the main video stream, drop all other streams.
+    ///
+    /// The output will be a single video stream.
+    #[arg(long)]
+    pub video_only: bool,
 }
 
 /// Sampling arguments.